move cluster score function and add sConj
